move intermediate classes only used for read/write to pf.io package
